Add owner attribute to volume summary
When retrieving a volume from cinder, we translate the object to a dict.
This dict now also includes an `owner` attribute read from
`os-vol-tenant-attr:tenant_id` - if it exists. It might not exist, if
the cinder policy does not allow the user to see that extension
attribute.

We need that attribute for boot-from-volume where the image-metadata is
read from the volume instead of the image. Image metadata contains an
owner attribute.

We will later on use the image's owner e.g. to chose a project to
download an image into.

(cherry picked from commit 63af5a06b6aeff0379408821306c627615835841)
(cherry picked from commit 154acce263f4ad12a911258b853c0b38b7c95626)

Change-Id: Ice5b753da4c0ee701a908d931a8b2a7a947f884a