Isolate request spec handling from _cold_migrate
The compat code for the request spec in _cold_migrate is
nearly as big as the rest of the method. This moves that to
another method and also slides the image meta handling within
the first condition where it's needed and renames the 'image'
variable to 'image_meta' to stop shadowing the 'image' module
import.

Change-Id: I83e1a1eba594897f4610e1f9074a24d79443a025