Decouple NVMe tests from os-brick
The existing tests for NVMe with Libvirt were using
APIs inside of os-brick rather than making sure that
we are calling os-brick with the right values.

This patch mocks up the factory function instead and
makes sure we call it with the right values, it also
adds a few more things to test if the root helper and
`device_scan_attempts` are passed down through to
os-brick as well.

Change-Id: I067f9e29b4db3bba0601041cf3603b9653a240b9