Remove stubs from VolumeAttachmentsSample API sample test
The amount of DB and compute service stubbing in these
functional tests is pretty gross and makes it harder to
maintain/extend them for new microversions which makes
it harder for new contributors to work with these kinds
of tests.

This removes the stubs and uses the CinderFixture. The
only new stub is dealing with detaching a volume with a
device tag since the fake driver does not track device
metadata on instances.

The API reference doc samples are regenerated using:

  tox -e api-samples VolumeAttachmentsSample

Change-Id: I04bca162c3a1d4fed7056385dfdca72c07bab9a5