Don't duplicate PlacementFixture in libvirt func tests
libvirt/base.py inherits from test_servers.ServersTestBase which
itself mixes in _IntegratedTestBase. _IntegratedTestBase starts
the PlacementFixture so we don't need to do so a second time.

If we do, then the things can get very weird during the test
cleanUp period with misleading exception and requests made to
placement when it seems like placement should already be gone
(because the inner fixture has alread been cleaned up, but not
the outer).

Change-Id: I057a07c8d0b880c8d09fc2e618ce1f7fc885beda