Fix the race in confirm resize func test
New confirm resize functional tests were added in
Ia69dace6d3b395fa113c1382e8eb52c2cc36119d. But they could be racy [1]
because the migration allocation is deleted in placement as the last
thing in the confirm_resize process. So it happens after the migration is
set to confirmed and the instance state is set to ACTIVE.

This patch changes the confirm resize tests to wait for the successful
instance action event as that is sent _after_ the migration allocation
is deleted.

[1] see the bug report

Closes-Bug: #1843433

Change-Id: Ibeb16ce16263c43bad9f148480bbebca413d8117