Skip querying resource request in revert_resize if no qos port
During revert_resize we can skip collecting port resource requests
from neutron if there is no port with allocation key in the binding
profile in the network info caches of the instance.

Note that we cannot do the same optimization in the MigrationTask as we
would like to use migration to heal missing port allocations. See more
in bug #1819923.

Change-Id: I72d0e28f3b9319e521243d7d0dc5dfa4feaf56f5
blueprint: support-move-ops-with-qos-ports