Make SRIOV computes non symmetric in func test
The FakeDriverWithPciResources creates 3 PFs with VFs for each compute
in a symmetric way. For example PF1 in every compute has a device name
'ens1'. When a server with SRIOV ports with resource request is booted
the PCI request of the server is updated to point to the PF dev name
that provides the bandwidth for the port. When such a server is migrated
this reference needs to be updated according to the destination host
resource allocation. As the compute were symmetric the update was a
no-op (e.g. from ens1 on the source to the ens1 on the destination
host). This prevent us to properly assert that the migration did the
update.

So this patch breaks the symmetry by appending the hostname to the PF
device names.

Change-Id: Id801d36f96b16708458db8549ce03abd308bc119
blueprint: support-move-ops-with-qos-ports