libvirt: stub logging of host capabilities
By default we capture the logging output stream at INFO level
for unit/functional tests and anything that hits the libvirt
Host.get_capabilities method logs the giant xml host capabilities
string that is defined in the fakelibvirt module. This can cause
subunit parsing to fail because there is too much output in the
buffer.

This change simply stubs out that particular logging call in the
FakeLibvirtFixture class which is used extensively in the more
detailed libvirt tests like LibvirtConnTestCase. If this turns out
to not be good enough we can try stubbing at a more global level
like in the base nova TestCase.

Change-Id: I38b350bce908005161ed7c3eb737f32076828a37
Related-Bug: #1813147