Use SpawnIsSynchronousFixture in reschedule functional tests
We see that in successful tests that are usually running some
kind of negative or reschedule scenario that they will dump
a stacktrace to the subunit output buffer because
_allocate_network_async will raise an error. Too many of these
can eventually blow up the subunit output buffer and cause
tests to fail.

To guard against this, we use the SpawnIsSynchronousFixture in
the base class of two known test classes [1][2] that show a
stacktace with _allocate_network_async in successful runs. As a
result, one of the tests has to be fixed due to improper stubbing.

[1] UnsupportedPortResourceRequestBasedSchedulingTest
[2] PortResourceRequestReSchedulingTest

Change-Id: Ib18ecaebbeccc9c9cea32ca8949175fcea0414ac
Related-Bug: #1813147