Remove redundancies from AggregateRequestFiltersTest.setUp
Change I9ab9d7d65378be564b3731b5227ede8cece71bef made
AggregateRequestFiltersTest extend ProviderUsageBaseTestCase
but left a lot of redundant setUp for fixtures and services
in it which might be contributing to test_fail_set_az failing
intermittently. My theory is that we're starting multiple API
and scheduler workers and when setting an AZ on an aggregate
the API will RPC cast to all schedulers to store that metadata
information in the scheduler process. Since we have more than
one scheduler process, and I'm not sure how stable the RPC cast
fanout capability is in the fake messaging driver, we could be
hitting a scheduler worker during the test that does not have
the AZ metadata and the AvailabilityZoneFilter filters out the
expected host.

This simply removes the redundant setUp. Even if this isn't the
root of the problem it at least rules it out.

Change-Id: I45107ff686456d5db91c615830fa443b3902ddfb
Partial-Bug: #1844174