Only allow one scheduler service in tests
There have been two recent issues [1][2] caused by starting multiple
instances of the same service in tests. This can cause races when the
services' (shared) state conflicts.

With this patch, the nexus of nova service starting,
nova.test.TestCase.start_service, is instrumented to keep track of how
many of each service we are running. If we try to run the scheduler
service more than once, we fail.

We could probably do the same thing for conductor, though that's less
important (for now) because conductor is stateless (for now).

[1] https://bugs.launchpad.net/nova/+bug/1844174
[2] https://review.opendev.org/#/c/681059/ (not a nova service, but same
class of problem)

Change-Id: I56d3cb17260dad8b88f03c0a7b9688efb3258d6f