Revert "nova shared storage: rbd is always shared storage"
This reverts commit 05b7f63a42e3512da6fe45d2e6639fb47ed8102b.

The _is_storage_shared_with method is specifically checking if the
instance directory is shared. It is not checking if the actual instance
disks are shared and as a result assumptions cannot be made based on the
value of images_type.

Closes-Bug: 1824858
Change-Id: I52293b6ce3e1ce77fa31b382d0067fb3bc68b21f
(cherry picked from commit 404932f82130445472837095b3ad9089c75e2660)