Refactor volume connection cleanup out of _post_live_migration
The _post_live_migration method is huge and hard to review/test/
maintain in isolation so we need to break down the complexity.
To that end this change refactors the source host volume connection
cleanup code into another method which also allows us to simplify
a couple of unit tests that target that specific set of code.
The error handling in that new method is tested by the functional
test class LiveMigrationCinderFailure.

Change-Id: Id0e8b1c32600d53382e5ac938e403258c80221a0