Trigger real BuildAbortException during migrate with bandwidth
The functional test
test_migrate_server_with_qos_port_pci_update_fail_not_reschedule used
fault injection to simulate that the
_update_pci_request_spec_with_allocated_interface_name call raises
BuildAbortException during migrating a instance with an SRIOV port
having resource request.

We can do a bit better. This patch simulates that neutron names the PF
device RP in placement in an unexpected way. Therefore the test does not
need to mock anything in nova to simulate the same failure.

Change-Id: Ib72f15eb62fd50ea7117df577478f555a8478724
blueprint: support-move-ops-with-qos-ports