Temporarily skip TestNovaMigrationsMySQL
There are several changes approved for Train feature freeze
which are still not merged yet nearly a week later because
of gate failures with bug 1823251. This change skips those
tests temporarily so we can land some code in time for Train
RC1 but this change should be reverted prior to RC1 after
all of those other changes are merged. Note that those
changes (vpmems and pcpu series) don't really have anything
to do with data model schema changes at this point so skipping
this test for now won't drop any coverage for what is going in,
and we aren't merging any other data model schema changes at
this point before RC1. Again, this is a temporary measure and
should be reverted within a few days.

Change-Id: I0564c613f30d321de33a1bd71bfe48c37e72ff80
Related-Bug: #1823251