nova-net: Migrate 'test_attach_interfaces' functional tests
This one is a little more involved and I'm not actually sure how it was
working before, since nova-net doesn't seem to support the 'show_port'
method. In short though, we're rewriting stuff to mock or wrap the
neutron implementation of nova's 'network' API. We also get to remove
some unnecessary mocks that the 'NeutronFixture' will now handle for us.

Change-Id: I15323a342d53cebd43f3a645528bc96489f19bc3
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>