hardware: Add TODO to remove '(un)pin_cpu_with_siblings'
With the newly introduced translation of 'hw:cpu_thread_policy' to
require or avoid the 'HW_CPU_HYPERTHREADING' trait, we no longer fake
non-SMT hosts by reserving the thread siblings of each core used by an
instance. As such, once we're in a world where all hosts have switched
to reporting PCPUs and this trait (or at least are capable of it), we
will no longer need the functionality these methods enable. Add a TODO
to remind us to remove them.

Change-Id: I22032f5a96c3ba791d3b44b83259b43349093c36
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>