Remove 'test_cold_migrate_with_physnet_fails' test
This never actually worked. What we wanted to do was test a setup where
we had two hosts, one with the necessary configuration needed for a
specific request and one without, and attempt to cold migrate the
instance from the former to the latter resulting in a fail. However,
because it's not possible to use different configuration for different
hosts, we were attempting to "break" the configuration on one host.
Unfortunately, the driver correctly detects this broken behavior,
resulting in an error message like so:

  ERROR [nova.compute.manager] Error updating resources for node test_compute1.
  Traceback (most recent call last):
    File "nova/compute/manager.py", line 8524, in _update_available_resource_for_node
      startup=startup)
    File "nova/compute/resource_tracker.py", line 867, in update_available_resource
      resources = self.driver.get_available_resource(nodename)
    File "nova/virt/libvirt/driver.py", line 7907, in get_available_resource
      numa_topology = self._get_host_numa_topology()
    File "nova/virt/libvirt/driver.py", line 7057, in _get_host_numa_topology
      physnet_affinities = _get_physnet_numa_affinity()
    File "nova/virt/libvirt/driver.py", line 7039, in _get_physnet_numa_affinity
      raise exception.InvalidNetworkNUMAAffinity(reason=msg)
  InvalidNetworkNUMAAffinity: Invalid NUMA network affinity configured: node 1 for
  physnet foo is not present in host affinity set {0: set([])}

There isn't really an alternative. We can't configure compute nodes
separately so that's ruled out and the virt driver will correctly detect
any other attempts to break the configuration. Since the test never
actually worked, the correct thing to do seems to be to remove it.

NOTE(stephenfin): This was backported to make the backport of change
I0322d872bdff68936033a6f5a54e8296a6fb3434 cleaner, but it also applies
here.

Change-Id: I14637d788205408dcf9a007d7727358c03033dcd
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>
(cherry picked from commit a4ca0b531cb3eabdeee6aab925220067ebd81172)