Reduce scope of 'path' query parameter to noVNC consoles
This is a partial revert of commit
9606c80402f6db20d62b689c58aa8f024183628a which added the 'path' query
parameter to work with noVNC v1.1.0. This broke all other console types
using websockify server (serial, spice) because the websockify server
itself doesn't know how to handle the 'path' query parameter. It is the
noVNC vnc_lite.html file which parses the 'path' variable and uses it
as the url to the websockify server. So, all other console types should
*not* be generating a console access url with a 'path' query parameter,
only noVNC.

Closes-Bug: #1845243

TODO(melwitt): Figure out how to test serial and/or spice console in
the gate

Change-Id: I9521f21a685edc44121d75bdf534c201fa87c2d7