[vmware] Add configurable reservations per hostgroup
This can be used to have different CPU/memory reservations for hosts
depending on the hostgroup they're in. We can also set defaults in
percent with this instead of only static values as supported by
CONF.reserved_host_memory_mb and CONF.reserved_host_cpus.

We also have to change the `os-hypervisors/details` call to adhere to
the reservations so both Placement and Nova report the same values. It's
important to us, because we get or used to get the quota capacity from
there.

I opted for changing the resources (for that single call, hence
copy.deepcopy) instead of subtracting the reserved values from the
totals afterwards, because if an attribute is unset on the `ComputeNode`
object, we cannot access it - not even with `getattr`. It just raises a
`NotImplementedError` because it cannot fetch the attribute from
somewhere (e.g. the DB). But since we need to subtract from the total,
we would have to access the attribute. Catching the
`NotImplementedError` didn't seem like a good option to me.

With subtracting the reserved resources, we have to make sure we do not
fall below zero - which could happen for Ironic nodes, where the memory
is set to 0. With us disallowing this, one of nova's assumptions failed
- that we don't collapse negative values to 0. But since that test only
worked with reservations, it couldn't detect that we still allow
negative values. We changed the test to test if instance memory can
still reduce the free memory to below zero.

NOTE: Setting these settings too differently per hostgroup for the same
cluster might result in placement API getting confused and scheduling
being impossible, as the `max_unit` is still set to the maximum per
cluster. In case of our HANA HVs placement would therefore always
consider the HANA HVs max unit. That should be no problem as smaller VMs
should always fit.

(cherry picked from commit 52504365ae43018a4aa70f8a11d7271aea18e62b)
(cherry picked from commit 28f4eb88a6170f27ca3d17dfd17c57f3cef3d227)

Change-Id: Ie4c9314bd19ec1565f3a380f74054fb1f84a0dcc
Incorporated-Change-Id: I191aedb0e3bba7698825771089cf134f320368ec
Incorporated-Change-Id: I1c92b193bf867db3a749b09bac57bde500c39e19