libvirt: Ignore DiskNotFound during update_available_resource
There was a previous attempt to fix this in
change Id687e11e235fd6c2f99bb647184310dfdce9a08d. However, there were 2
problems with the previous fix:

1. The handling of missing volumes and disks, while typically having the
   same cause, was inconsistent.

2. It failed to consider the very wide race opportunity in
   _get_disk_over_committed_size_total between initially fetching the
   instance list from the DB and later getting disk sizes.

Because _get_disk_over_committed_size_total() can be a very long
operation, we found that we were reliably hitting this race in CI.
It might be possible to fix the race, but this would add unnecessary
complication to code which isn't critical. It's far more robust just to
log it and ignore it, which is also consistent with the handling of
missing volumes.

Closes-Bug: #1774249

Change-Id: I48719c02713113a41176b8f5cc3c5831f1284a39
(cherry picked from commit 6198f317be549e6d2bd324a48f226b379556e945)