test cleanup: Make base TestCase subclass oslotest
oslotest.base.BaseTestCase already sets up many of the fixtures we want
from nova.test.TestCase, so inherit from it and remove those setups.

Note that oslotest has a logger fixture, but nova's is significantly
different, so we preserve the latter. Fortunately, oslotest provides a
granny switch to disable its logger fixture -- the OS_LOG_CAPTURE env
var -- so we explicitly do that, even though it's not on in our tox.ini,
in case we accidentally run in an env where that's enabled.

Change-Id: Ic61c3db824ec811336423bd956f2f2f03e02d21e