Add Selection.availability_zone field
This adds an availability_zone field to the Selection object,
using the same type and nullable value as the same field in the
Instance object. This will be used to store the service_host
AZ to pass from the superconductor layer to the compute and cell
conductor layer to avoid an up-call to get the host AZ information
from the API DB during a reschedule.

Note that the field is nullable because a host may not be in an
AZ and CONF.default_availability_zone can technically be set to
None though it defaults to "nova".

Change-Id: Ia50c5f4dd2204f1cafa669097d1e744479c4d8c8
Related-Bug: #1781286