Add functional regression test for build part of bug 1781286
This adds a functional regression test to recreate bug 1781286
where rescheduling from a build failure where the cell conductor
does not have access to the API DB fails when trying to get
availability zone information about a selected alternate host.
The test is a bit tricky in that it has to stub out the AZ query
to fail but only after we hit the compute service and then remove
that stub so we can use the API again.

Note that change I6bfa6fa767403fb936a6ae340b8687eb161732fc handles
the error in conductor and puts the instance into ERROR state. The
bug is not yet resolved though since we should be able to avoid
the up-call by stashing the alternate host availability zone on the
Selection object during scheduling and use that later during the
reschedule. Once that is working the functional test will be updated.

Change-Id: I62179d6b93ea1a23c4906477ee19b422bfcb72a2
Related-Bug: #1781286