Set Instance AZ from Selection AZ during build reschedule
This builds on change Ia50c5f4dd2204f1cafa669097d1e744479c4d8c8
to use the Selection.availability_zone value when rescheduling
during initial server create so that the cell conductor does not
have to make an up-call to the aggregates table in the API DB
which will fail if the cell conductor is not configured to use
the API DB.

The functional test added in I62179d6b93ea1a23c4906477ee19b422bfcb72a2
is updated to show the failure is gone and we get the AZ from the
Selection object during the reschedule.

For the case that the availability_zone field is not in the Selection
object, test_build_reschedule_get_az_error still covers that.

Change-Id: I1f1c25cb4de924a1d6c3a979b758efd736bdbff0
Partial-Bug: #1781286