Add boot from volume functional test with a huge request
While discussing rate limiting, or the lack thereof, in the API
the point came up that it's possible to create multiple servers
with a large number of BDMs in a single request and it's not
rate limited nor is there any optimization in the API for not
GETing the same image in the request more than once.

This adds a functional test to demonstrate by wrapping the image
API GET calls and asserting they are not cached within the request.

Related-Bug: #1846777

Change-Id: I1233fbbaebd1750f572432bf74759b94408a17e3