fixtures: Store 'device_id' when creating port in NeutronFixture
When creating a port, we send a bit of information to neutron, namely,
the requested network, the tenant (project) ID and, crucially, the
device ID. The 'NeutronFixture' should be setting these fields on the
created port but is not. This hasn't been an issue to date but will when
we add subnet support, so fix this now.

Change-Id: Ia842854c12cd2d2ed630d1e0f2bb691859592cd9
Signed-off-by: Stephen Finucane <sfinucan@redhat.com>