Cache image GETs for multi-create/multi-BDM requests
This builds on change I1233fbbaebd1750f572432bf74759b94408a17e3 which
added a functional test to create 10 servers with 255 BDMs each using
the two images. That scenario results in 2551 calls to the image service
to get the images. While an unrealistic scenario, it shows how
inefficient this can be even in a simple multi-create request where
each server has one or more BDMs that are using one or more of the
same images.

This change adds a simple cache that is used while processing the list
of instances to create and the list of BDMs per instance that are using
a source_type of 'image' to validate the image. This drops the image GET
calls in the functional test from 2551 to 3.

The functional test verifies the cache is being used properly. Existing
unit tests are updated for the method signature changes. Note that some
API controller tests were mocking a bit too deep and ignoring all of the
parameters to the _validate_bdm mock so rather than continue that pattern
the assertions are just changed to assert_called_once() since the tests
clearly don't care about the actual parameters passed to that method.

Change-Id: Ibd51d6fc2d4c9fd142ba17f2f87214bb62f72440
Closes-Bug: #1846777